refactor(agent): remove legacy code and consolidate abstractions
Major cleanup removing deprecated code and unnecessary abstractions:

Legacy Code Removed:
- Remove legacy UsageSummary migration (AgentState.ts: -115 lines)
  - Deleted migrateLegacyUsageSummary() and related type guards
  - Removed deprecated schema fields (usageSummary, nativeUsage, provider)
- Remove legacy snapshot migration (RunUsageAccumulator.ts: -130 lines)
  - Deleted migrateLegacySnapshot() and LegacyNativeUsageSnapshot type
  - Removed deprecated UsageSummary and NativeUsagePayload exports
  - Cleaned up schema to remove legacy totalToolUseTokens alias

Abstractions Consolidated:
- Merge FollowUpQueue into ToolUseFollowUpQueue.ts
  - Deleted separate FollowUpQueue.ts file
  - Single file now contains both classes
- Delete ToolUseSessionManager.ts re-export file
  - Updated imports to use ToolUseSnapshotCache directly
- Simplify runStateSchemas.ts
  - Removed unused isBaseRunState type guard
  - Removed unnecessary re-exports

Code Quality:
- Removed redundant defensive copies in RunUsageAccumulator
  - getTotals() now returns direct reference
  - getNormalizedSnapshots() returns readonly array
- Simplified fromJSON() methods without legacy fallbacks

Net reduction: ~320 lines of code
