+++ src/agent/core/RunUsageAccumulator.ts
@@ -5,30 +5,7 @@ import { z } from 'zod';
 import {
   NormalizedUsageSchema,
   type NormalizedUsage,
-  type UsageProvider,
 } from '@agent/types/NormalizedUsage';
-
-// Re-export for backwards compatibility (used in AgentState legacy schema)
-import type {
-  AnthropicAPIResponseUsage,
-  OpenAIAPIResponseUsage,
-  NativeUsagePayload,
-} from './ResponseUsage';
-
-export type { NativeUsagePayload };
-
-/** @deprecated Used only for legacy JSON deserialization */
-export type UsageSummary =
-  | OpenAIAPIResponseUsage
-  | AnthropicAPIResponseUsage
-  | null;
-
-/** @deprecated Legacy snapshot format - used only for migration */
-interface LegacyNativeUsageSnapshot {
-  round: number;
-  provider: string;
-  payload: unknown;
-}
 
 /** Schema for run usage totals */
 export const RunUsageTotalsSchema = z.object({
@@ -55,13 +32,8 @@ export type NormalizedUsageSnapshot = z.infer<
 
 /** Schema for RunUsageAccumulator JSON serialization */
 export const RunUsageAccumulatorJSONSchema = z.object({
-  totals: RunUsageTotalsSchema.partial().extend({
-    /** @deprecated Legacy field name */
-    totalToolUseTokens: z.number().optional(),
-  }),
+  totals: RunUsageTotalsSchema.partial(),
   normalizedSnapshots: z.array(NormalizedUsageSnapshotSchema).optional(),
-  /** @deprecated Legacy format - ignored on load */
-  snapshots: z.array(z.unknown()).optional(),
 });
 export type RunUsageAccumulatorJSON = z.infer<
   typeof RunUsageAccumulatorJSONSchema
@@ -83,12 +55,8 @@ export class RunUsageAccumulator {
   private totals: RunUsageTotals = { ...DEFAULT_TOTALS };
   private readonly normalizedSnapshots: NormalizedUsageSnapshot[] = [];
 
-  /**
-   * Records normalized usage from a model response.
-   */
   recordNormalizedUsage(round: number, usage: NormalizedUsage): void {
     if (this.totals.firstInputTokens === 0) {
-      // Include cached tokens in first input count (Anthropic reports them separately)
       this.totals.firstInputTokens =
         usage.inputTokens +
         (usage.cachedInputTokens ?? 0) +
@@ -108,7 +76,7 @@ export class RunUsageAccumulator {
   }
 
   merge(other: RunUsageAccumulator): void {
-    const otherTotals = other.getTotals();
+    const otherTotals = other.totals;
     if (this.totals.firstInputTokens === 0) {
       this.totals.firstInputTokens = otherTotals.firstInputTokens;
     }
@@ -125,21 +93,24 @@ export class RunUsageAccumulator {
       otherTotals.totalToolUsePromptTokens;
     this.totals.totalServerToolRequests += otherTotals.totalServerToolRequests;
 
-    this.normalizedSnapshots.push(...other.getNormalizedSnapshots());
+    this.normalizedSnapshots.push(...other.normalizedSnapshots);
   }
 
   getTotals(): RunUsageTotals {
-    return { ...this.totals };
+    return this.totals;
   }
 
-  getNormalizedSnapshots(): NormalizedUsageSnapshot[] {
-    return [...this.normalizedSnapshots];
+  getNormalizedSnapshots(): readonly NormalizedUsageSnapshot[] {
+    return this.normalizedSnapshots;
   }
 
   toJSON(): RunUsageAccumulatorJSON {
     return {
-      totals: this.getTotals(),
-      normalizedSnapshots: this.getNormalizedSnapshots(),
+      totals: this.totals,
+      normalizedSnapshots:
+        this.normalizedSnapshots.length > 0
+          ? this.normalizedSnapshots
+          : undefined,
     };
   }
 
@@ -158,117 +129,14 @@ export class RunUsageAccumulator {
       totalCacheReadInputTokens: t.totalCacheReadInputTokens ?? 0,
       totalCacheCreationInputTokens: t.totalCacheCreationInputTokens ?? 0,
       totalReasoningTokens: t.totalReasoningTokens ?? 0,
-      // Handle legacy field name
-      totalToolUsePromptTokens:
-        t.totalToolUsePromptTokens ?? t.totalToolUseTokens ?? 0,
+      totalToolUsePromptTokens: t.totalToolUsePromptTokens ?? 0,
       totalServerToolRequests: t.totalServerToolRequests ?? 0,
     };
 
     if (json.normalizedSnapshots) {
       acc.normalizedSnapshots.push(...json.normalizedSnapshots);
     }
-    // Migrate legacy snapshots format to normalized format
-    else if (json.snapshots && Array.isArray(json.snapshots)) {
-      for (const snap of json.snapshots as LegacyNativeUsageSnapshot[]) {
-        const migrated = migrateLegacySnapshot(snap);
-        if (migrated) {
-          acc.normalizedSnapshots.push(migrated);
-        }
-      }
-    }
 
     return acc;
   }
-}
-
-/**
- * Migrates a legacy native usage snapshot to normalized format.
- * Extracts tokens from native payload; cost cannot be recovered (depends on pricing).
- */
-function migrateLegacySnapshot(
-  legacy: LegacyNativeUsageSnapshot,
-): NormalizedUsageSnapshot | null {
-  if (!legacy.payload || typeof legacy.payload !== 'object') {
-    return null;
-  }
-
-  const payload = legacy.payload as Record<string, unknown>;
-  const provider = (legacy.provider || 'unknown') as UsageProvider;
-
-  // Try to detect payload format and extract tokens
-  // Anthropic format: input_tokens, output_tokens
-  if ('input_tokens' in payload) {
-    return {
-      round: legacy.round,
-      usage: {
-        inputTokens: (payload.input_tokens as number) ?? 0,
-        outputTokens: (payload.output_tokens as number) ?? 0,
-        cost: 0, // Cannot recover - depends on pricing at time of request
-        responseTimeMs: 0,
-        provider,
-        cachedInputTokens:
-          (payload.cache_read_input_tokens as number) ?? undefined,
-        cacheCreationTokens:
-          (payload.cache_creation_input_tokens as number) ?? undefined,
-        _native: payload,
-      },
-    };
-  }
-
-  // OpenAI format: prompt_tokens, completion_tokens
-  if ('prompt_tokens' in payload) {
-    const details = payload.prompt_tokens_details as
-      | Record<string, unknown>
-      | undefined;
-    const completionDetails = payload.completion_tokens_details as
-      | Record<string, unknown>
-      | undefined;
-    return {
-      round: legacy.round,
-      usage: {
-        inputTokens: (payload.prompt_tokens as number) ?? 0,
-        outputTokens: (payload.completion_tokens as number) ?? 0,
-        cost: 0, // Cannot recover - depends on pricing at time of request
-        responseTimeMs: 0,
-        provider,
-        cachedInputTokens: (details?.cached_tokens as number) ?? undefined,
-        reasoningTokens:
-          (completionDetails?.reasoning_tokens as number) ?? undefined,
-        _native: payload,
-      },
-    };
-  }
-
-  // Google format: promptTokenCount, candidatesTokenCount
-  if ('promptTokenCount' in payload) {
-    return {
-      round: legacy.round,
-      usage: {
-        inputTokens: (payload.promptTokenCount as number) ?? 0,
-        outputTokens: (payload.candidatesTokenCount as number) ?? 0,
-        cost: 0, // Cannot recover - depends on pricing at time of request
-        responseTimeMs: 0,
-        provider,
-        cachedInputTokens:
-          (payload.cachedContentTokenCount as number) ?? undefined,
-        reasoningTokens: (payload.thoughtsTokenCount as number) ?? undefined,
-        toolUsePromptTokens:
-          (payload.toolUsePromptTokenCount as number) ?? undefined,
-        _native: payload,
-      },
-    };
-  }
-
-  // Unknown format - create minimal entry preserving native payload
-  return {
-    round: legacy.round,
-    usage: {
-      inputTokens: 0,
-      outputTokens: 0,
-      cost: 0,
-      responseTimeMs: 0,
-      provider,
-      _native: payload,
-    },
-  };
 }
